Move svg and CTACard to separate files, rename .jsx to .js and use className instead of class attributes
