--- client/my-sites/google-my-business/select-business-type/index.jsx
@@ -1,657 +1,0 @@
-/**
- * External dependencies
- *
- * @format
- */
-import React, { Component } from 'react';
-import { connect } from 'react-redux';
-import { localize } from 'i18n-calypso';
-import Gridicon from 'gridicons';
-
-/**
- * Internal dependencies
- */
-import HeaderCake from 'components/header-cake';
-import Card from 'components/card';
-import Button from 'components/button';
-
-const CTACard = props => (
-	<Card className="select-business-type__cta-card">
-		<div class="select-business-type__cta-card__main">
-			<h2>{ props.headerText }</h2>
-			<p>{ props.mainText }</p>
-		</div>
-		<div class="select-business-type__cta-card__button-container">
-			<Button primary={ props.buttonPrimary }>
-				{ props.buttonText } <Gridicon icon={ props.buttonIcon } />
-			</Button>
-		</div>
-	</Card>
-);
-
-class SelectBusinessType extends Component {
-	render() {
-		const { translate } = this.props;
-
-		return (
-			<div class="select-business-type">
-				<HeaderCake isCompact={ false } alwaysShowActionText={ false }>
-					{ translate( 'Google My Business' ) }
-				</HeaderCake>
-				<Card className="select-business-type__explanation">
-					<div class="select-business-type__explanation__main">
-						<h1>{ translate( 'Which type of business are you?' ) }</h1>
-						<h2>
-							{ translate(
-								'Google My Business lists your local business on Google Search and Google Maps. ' +
-									'It works for businesses that have a physical location or serve a local area.'
-							) }
-						</h2>
-					</div>
-					<svg
-						width="315px"
-						height="129px"
-						viewBox="0 0 315 129"
-						version="1.1"
-						xmlns="http://www.w3.org/2000/svg"
-					>
-						<g id="Assets" stroke="none" stroke-width="1" fill="none" fill-rule="evenodd">
-							<g id="gmb-illustration">
-								<g id="Page-1">
-									<path
-										d="M156.631256,50.2270682 C157.434348,50.0470007 158.237441,50.0143752 159.040533,50.0193945 C159.843626,50.0306879 160.646719,50.0739794 161.449811,50.2258134 C161.503196,50.235852 161.539173,50.2910644 161.529308,50.3487864 C161.522345,50.3939601 161.489269,50.427213 161.449811,50.4353694 C160.646719,50.5872034 159.843626,50.6304949 159.040533,50.6411609 C158.237441,50.6468076 157.434348,50.6135547 156.631256,50.4341146 C156.578451,50.4221937 156.544795,50.366354 156.55524,50.3092594 C156.563364,50.2665953 156.594699,50.2352246 156.631256,50.2270682"
-										id="Fill-41"
-										fill="#C7D6E1"
-									/>
-									<path
-										d="M179.653031,68.6978688 C180.456123,68.5184287 181.259216,68.4851758 182.062309,68.4908225 C182.865401,68.5014885 183.668494,68.54478 184.471586,68.6972414 C184.524971,68.70728 184.560948,68.761865 184.551083,68.819587 C184.54412,68.8647607 184.511045,68.898641 184.471586,68.90617 C183.668494,69.0580039 182.865401,69.1012955 182.062309,69.1119615 C181.259216,69.1176082 180.456123,69.0849827 179.653031,68.9049151 C179.600226,68.8929943 179.566571,68.8371545 179.577016,68.7800599 C179.585139,68.7373958 179.616474,68.7060252 179.653031,68.6978688"
-										id="Fill-43"
-										fill="#C7D6E1"
-									/>
-									<path
-										d="M121.831002,58.2578396 C122.634094,58.0777721 123.437187,58.0451466 124.240279,58.0507933 C125.043372,58.0614593 125.846465,58.1047508 126.649557,58.2565848 C126.702942,58.2666234 126.738919,58.3218358 126.729054,58.3795578 C126.722091,58.4247315 126.689015,58.4579844 126.649557,58.4661408 C125.846465,58.6179748 125.043372,58.6612663 124.240279,58.6719323 C123.437187,58.677579 122.634094,58.6443261 121.831002,58.464886 C121.778197,58.4529651 121.743961,58.3971253 121.754986,58.3400307 C121.76311,58.2973666 121.793864,58.265996 121.831002,58.2578396"
-										id="Fill-45"
-										fill="#C7D6E1"
-									/>
-									<polygon
-										id="Fill-172"
-										fill="#7ED2F1"
-										points="229.092259 111.132965 288.052834 111.132965 288.052834 60.9669502 229.092259 60.9669502"
-									/>
-									<path
-										d="M235.599166,72.1109046 L235.599166,89.3715929 L261.851706,89.3715929 L261.851706,65.270691 L242.439379,65.270691 C238.661246,65.270691 235.599166,68.3333519 235.599166,72.1109046"
-										id="Fill-173"
-										fill="#C6E8F4"
-									/>
-									<polygon
-										id="Fill-174"
-										fill="#C6E8F4"
-										points="267.016334 108.307635 282.509522 108.307635 282.509522 65.270807 267.016334 65.270807"
-									/>
-									<polygon
-										id="Fill-175"
-										fill="#FFFFFF"
-										points="269.268939 84.3396143 280.464654 84.3396143 280.464654 68.0375307 269.268939 68.0375307"
-									/>
-									<path
-										d="M280.803369,31.1620018 L280.872061,31.1620018 L294.605292,51.8584642 L294.598677,51.8584642 L294.598677,55.0656122 C294.598677,58.22924 292.102939,60.8247842 288.942213,60.9495421 C288.659622,60.9611475 288.363684,60.9669502 288.05266,60.9669502 L229.092085,60.9669502 L229.092085,51.8584642 L229.098119,31.1620018 L229.098352,25.6967361 L280.47887,25.6967361 L280.803369,31.1620018 Z"
-										id="Combined-Shape"
-										fill="#07A8DC"
-									/>
-									<path
-										d="M281.317185,93.2510988 L298.532032,93.2510988"
-										id="Fill-179"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M286.539608,110.552638 L293.309609,110.552638"
-										id="Fill-183"
-										fill="#FFFFFF"
-									/>
-									<polygon
-										id="Fill-193"
-										fill="#C6E8F4"
-										points="170.783906 111.065654 229.026688 111.065654 229.026688 26.4902543 170.783906 26.4902543"
-									/>
-									<path
-										d="M206.641236,83.3094623 L180.703202,83.3094623 C178.096633,83.3094623 175.983872,81.1967021 175.983872,78.5901327 L175.983872,52.6520984 L206.641236,52.6520984 L206.641236,83.3094623 Z"
-										id="Fill-194"
-										fill="#7ED2F1"
-									/>
-									<polygon
-										id="Fill-195"
-										fill="#FFFFFF"
-										points="175.983698 49.5571165 224.339274 49.5571165 224.339274 39.6925397 175.983698 39.6925397"
-									/>
-									<polygon
-										id="Fill-196"
-										fill="#7ED2F1"
-										points="209.639313 88.822194 224.339274 88.822194 224.339274 52.6522725 209.639313 52.6522725"
-									/>
-									<path
-										d="M170.56793,88.8220199 L229.92599,88.8220199"
-										id="Fill-197"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M170.56793,88.4349803 C173.041037,88.1639946 175.514145,88.2591588 177.987832,88.229565 L185.407735,88.2051937 C190.35395,88.199391 195.300745,88.1489076 200.24696,88.1814027 L215.086765,88.2301453 C220.03298,88.227244 224.979775,88.2933947 229.92599,88.62821 L229.92599,89.0152496 C224.979775,89.3500649 220.03298,89.4167959 215.086765,89.4138945 L200.24696,89.4626371 C195.300745,89.4951322 190.35395,89.4446488 185.407735,89.4382658 L177.987832,89.4138945 C175.514145,89.3843008 173.041037,89.4800452 170.56793,89.2084792 L170.56793,88.4349803 Z"
-										id="Fill-198"
-										fill="#0383BF"
-									/>
-									<path
-										d="M229.02663,94.2376726 L208.418369,94.2376726"
-										id="Fill-199"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M229.02663,94.6247122 C228.167832,94.9009203 227.309614,94.7970521 226.450815,94.8301275 L223.87442,94.8544988 C222.157403,94.8637831 220.439806,94.9026611 218.72279,94.8869938 C217.005193,94.8893149 215.287596,94.8655239 213.570579,94.8365104 C211.852982,94.8202629 210.135385,94.7784835 208.418369,94.4314825 L208.418369,94.0444429 C210.135385,93.6968617 211.852982,93.6556626 213.570579,93.639415 C215.287596,93.6098213 217.005193,93.5860302 218.72279,93.5883513 C220.439806,93.5726841 222.157403,93.6115621 223.87442,93.6208464 L226.450815,93.645798 C227.309614,93.6782931 228.167832,93.5744249 229.02663,93.8512133 L229.02663,94.6247122 Z"
-										id="Fill-200"
-										fill="#0383BF"
-									/>
-									<path
-										d="M208.418311,98.8798264 L229.026572,98.8798264"
-										id="Fill-201"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M208.418311,98.4930769 C209.277109,98.2168688 210.135327,98.320737 210.994126,98.2876616 L213.570521,98.2632903 C215.287538,98.2534257 217.005135,98.215128 218.722732,98.2307952 C220.439748,98.2278939 222.157345,98.2522652 223.874362,98.2812787 C225.591959,98.2975262 227.309556,98.3387253 229.026572,98.6863066 L229.026572,99.0733462 C227.309556,99.4209274 225.591959,99.4621265 223.874362,99.4783741 C222.157345,99.5073875 220.439748,99.5317588 218.722732,99.5288575 C217.005135,99.545105 215.287538,99.506227 213.570521,99.4963624 L210.994126,99.4719911 C210.135327,99.439496 209.277109,99.542784 208.418311,99.2665758 L208.418311,98.4930769 Z"
-										id="Fill-202"
-										fill="#0383BF"
-									/>
-									<path
-										d="M208.418311,103.52198 L229.026572,103.52198"
-										id="Fill-203"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M208.418311,103.135231 C209.277109,102.859023 210.135327,102.962891 210.994126,102.929815 L213.570521,102.905444 C215.287538,102.89558 217.005135,102.857282 218.722732,102.872949 C220.439748,102.870048 222.157345,102.894419 223.874362,102.923432 C225.591959,102.93968 227.309556,102.980879 229.026572,103.32846 L229.026572,103.7155 C227.309556,104.063081 225.591959,104.10428 223.874362,104.120528 C222.157345,104.149541 220.439748,104.173913 218.722732,104.171011 C217.005135,104.187259 215.287538,104.148381 213.570521,104.138516 L210.994126,104.114145 C210.135327,104.08165 209.277109,104.184938 208.418311,103.90873 L208.418311,103.135231 Z"
-										id="Fill-204"
-										fill="#0383BF"
-									/>
-									<path
-										d="M186.978407,69.2200613 L184.299304,67.9161964 C186.262355,62.2429042 190.481493,62.0868118 190.481493,62.0868118 L192.013984,63.1156291 L193.569686,62.0868118 C197.089599,62.960117 198.326152,67.307494 198.326152,68.6949177 L195.693471,69.4016856 L195.918035,75.4689806 L186.651136,75.4689806 L186.978407,69.2200613 Z"
-										id="Fill-205"
-										fill="#FFFFFF"
-									/>
-									<polygon
-										id="Fill-206"
-										fill="#005080"
-										points="102.170958 111.027763 102.170958 41.8927465 116.827398 17.8342042 134.249401 43.5523165 170.856846 43.8285246 170.856846 111.027763"
-									/>
-									<polygon
-										id="Fill-207"
-										fill="#07A8DC"
-										points="112.326424 93.2473271 120.898741 93.2473271 120.898741 70.7369431 112.326424 70.7369431"
-									/>
-									<polygon
-										id="Fill-208"
-										fill="#07A8DC"
-										points="137.552932 68.0181497 137.552932 111.227897 149.712473 111.227897 161.872015 111.227897 161.872015 68.0181497"
-									/>
-									<path
-										d="M116.612524,35.7474633 C115.868619,35.7567476 112.326076,36.0317952 112.326076,40.4226924 L112.326076,55.6541793 L116.612524,55.6541793 L120.898973,55.6541793 L120.898973,40.4226924 C120.898973,36.0317952 117.356429,35.7567476 116.612524,35.7474633"
-										id="Fill-209"
-										fill="#07A8DC"
-									/>
-									<polygon
-										id="Fill-210"
-										fill="#7ED2F1"
-										points="142.037369 111.22813 157.38723 111.22813 157.38723 80.2313082 142.037369 80.2313082"
-									/>
-									<polygon
-										id="Fill-211"
-										fill="#07A8DC"
-										points="95.5757921 43.5785446 117.106101 9.54517442 126.476289 23.2511335 170.85644 23.2279227 170.85644 43.8669384 133.947835 43.7526254 117.106101 18.0345132 101.589122 43.5785446"
-									/>
-									<path
-										d="M200.431949,26.4900802 L233.212518,26.4900802 C228.08584,22.2831284 224.668054,0.377965178 224.668054,0.377965178 L200.431949,0.0222601444 L200.431949,0.0199390675 L200.354774,0.021099606 L200.277598,0.0199390675 L200.277598,0.0222601444 L176.042073,0.377965178 C176.042073,0.377965178 172.624288,22.2831284 167.497029,26.4900802 L200.277598,26.4900802"
-										id="Fill-212"
-										fill="#005080"
-									/>
-									<path
-										d="M137.146047,69.7736962 C137.146047,62.8174288 142.785684,57.1772119 149.742531,57.1772119 C156.698799,57.1772119 162.338435,62.8174288 162.338435,69.7736962 L162.338435,73.2698183 L137.146047,73.2698183 L137.146047,69.7736962 Z"
-										id="Fill-213"
-										fill="#C6E8F4"
-									/>
-									<path
-										d="M162.924972,111.300605 C160.72401,111.520527 158.523049,111.518786 156.322088,111.531552 L149.719205,111.57043 L143.116321,111.531552 C140.91594,111.518786 138.714979,111.521107 136.514018,111.300605 L136.514018,111.155538 C138.714979,110.935036 140.91594,110.937357 143.116321,110.924591 L149.719205,110.885713 L156.322088,110.924591 C158.523049,110.937357 160.72401,110.935616 162.924972,111.155538 L162.924972,111.300605 Z"
-										id="Fill-214"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M152.850744,65.7152352 C153.222696,65.1906719 153.286526,64.0388374 152.528694,63.2827467 C151.478987,62.2342002 149.741661,61.1380716 149.741661,61.1380716 C149.741661,61.1380716 149.506652,61.7572189 148.313618,62.0839104 C146.619813,62.5475455 146.95927,63.6935773 146.95927,63.6935773 C146.092928,64.1728796 145.725618,65.0722969 146.382482,65.7152352"
-										id="Fill-215"
-										fill="#07A8DC"
-									/>
-									<path
-										d="M145.990046,65.9039388 L146.880179,69.1476437 C147.034531,69.7087641 147.544007,70.0975445 148.126017,70.0975445 L151.249607,70.0975445 C151.829876,70.0975445 152.338772,69.7110852 152.494284,69.1522859 L153.398924,65.9039388"
-										id="Fill-216"
-										fill="#7ED2F1"
-									/>
-									<path
-										d="M146.416486,65.9308053 C145.987087,65.5843845 145.761943,64.8903825 146.040472,64.3083725 C146.312618,63.7321652 146.874899,63.4060539 147.402944,63.2366152 C147.938532,63.0613739 148.508937,63.0155327 149.025377,62.9766546 C149.542977,62.9354555 150.055935,62.8426124 150.581659,62.753251 L150.635624,62.9186277 C150.147617,63.1600197 149.617832,63.3091489 149.077021,63.3985104 C148.53621,63.4716243 148.036018,63.5250091 147.559617,63.7031517 C147.090759,63.8598244 146.649174,64.1128218 146.433894,64.5016022 C146.212232,64.8869009 146.255171,65.3981181 146.484958,65.8774205 L146.416486,65.9308053 Z"
-										id="Fill-217"
-										fill="#7ED2F1"
-									/>
-									<path
-										d="M145.453994,65.5899551 C145.802736,65.451851 146.150897,65.503495 146.499639,65.4872475 L147.545284,65.4750618 C148.242768,65.4704196 148.939671,65.4506905 149.637155,65.4588143 C150.334638,65.4576537 151.031541,65.4692591 151.728445,65.4843461 C152.425928,65.4924699 153.122832,65.5127793 153.820315,65.6868601 L153.820315,65.8800897 C153.122832,66.0535902 152.425928,66.0744799 151.728445,66.0826037 C151.031541,66.0971104 150.334638,66.1092961 149.637155,66.1081355 C148.939671,66.1162593 148.242768,66.0965301 147.545284,66.091888 L146.499639,66.0797023 C146.150897,66.0628745 145.802736,66.1150988 145.453994,65.9769947 L145.453994,65.5899551 Z"
-										id="Fill-218"
-										fill="#005080"
-									/>
-									<path
-										d="M152.473336,62.9036568 C152.548771,63.2309286 152.422273,63.6052023 152.193647,63.8593602 C151.966761,64.1204813 151.682429,64.3119702 151.383591,64.4448518 C150.77895,64.690886 150.147037,64.8231874 149.516284,64.7454313 L149.497716,64.5725111 C150.079726,64.358972 150.665798,64.2585854 151.178176,63.9974643 C151.433494,63.8756077 151.676627,63.7375036 151.875659,63.5570399 C152.080494,63.3777367 152.231364,63.1717411 152.386876,62.8932119 L152.473336,62.9036568 Z"
-										id="Fill-219"
-										fill="#7ED2F1"
-									/>
-									<path
-										d="M141.344817,73.1258535 C141.344817,74.2852314 140.404781,75.2252675 139.245403,75.2252675 C138.086025,75.2252675 137.145989,74.2852314 137.145989,73.1258535 C137.145989,71.9658953 138.086025,71.0264394 139.245403,71.0264394 C140.404781,71.0264394 141.344817,71.9658953 141.344817,73.1258535"
-										id="Fill-220"
-										fill="#C6E8F4"
-									/>
-									<path
-										d="M145.543529,73.1258535 C145.543529,74.2852314 144.603493,75.2252675 143.444115,75.2252675 C142.284737,75.2252675 141.344701,74.2852314 141.344701,73.1258535 C141.344701,71.9658953 142.284737,71.0264394 143.444115,71.0264394 C144.603493,71.0264394 145.543529,71.9658953 145.543529,73.1258535"
-										id="Fill-221"
-										fill="#C6E8F4"
-									/>
-									<path
-										d="M149.742241,73.1258535 C149.742241,74.2852314 148.802205,75.2252675 147.642827,75.2252675 C146.483449,75.2252675 145.543413,74.2852314 145.543413,73.1258535 C145.543413,71.9658953 146.483449,71.0264394 147.642827,71.0264394 C148.802205,71.0264394 149.742241,71.9658953 149.742241,73.1258535"
-										id="Fill-222"
-										fill="#C6E8F4"
-									/>
-									<path
-										d="M153.941011,73.1258535 C153.941011,74.2852314 153.000975,75.2252675 151.841597,75.2252675 C150.682219,75.2252675 149.742183,74.2852314 149.742183,73.1258535 C149.742183,71.9658953 150.682219,71.0264394 151.841597,71.0264394 C153.000975,71.0264394 153.941011,71.9658953 153.941011,73.1258535"
-										id="Fill-223"
-										fill="#C6E8F4"
-									/>
-									<path
-										d="M158.139723,73.1258535 C158.139723,74.2852314 157.199687,75.2252675 156.040309,75.2252675 C154.880931,75.2252675 153.940895,74.2852314 153.940895,73.1258535 C153.940895,71.9658953 154.880931,71.0264394 156.040309,71.0264394 C157.199687,71.0264394 158.139723,71.9658953 158.139723,73.1258535"
-										id="Fill-224"
-										fill="#C6E8F4"
-									/>
-									<path
-										d="M162.338203,73.1258535 C162.338203,74.2852314 161.398167,75.2252675 160.238789,75.2252675 C159.079411,75.2252675 158.139956,74.2852314 158.139956,73.1258535 C158.139956,71.9658953 159.079411,71.0264394 160.238789,71.0264394 C161.398167,71.0264394 162.338203,71.9658953 162.338203,73.1258535"
-										id="Fill-225"
-										fill="#C6E8F4"
-									/>
-									<path
-										d="M0.0838624801,110.651806 L19.7554049,110.456255 L39.4273694,110.362252 L78.7708763,110.184689 L157.45789,110.111575 L236.144904,110.205579 L275.488411,110.333819 L295.159953,110.404611 L314.831918,110.535752 C314.925182,110.536333 315.000299,110.640781 314.999877,110.769021 C314.999455,110.896099 314.924338,110.999387 314.831918,110.999968 L295.159953,111.131689 L275.488411,111.201901 L236.144904,111.330141 L157.45789,111.424725 L78.7708763,111.351031 L39.4273694,111.173468 L19.7554049,111.079465 L0.0838624801,110.883914 C0.0370194456,110.883334 -0.000539383848,110.831109 -0.000117374528,110.76728 C0.000304634792,110.70345 0.0378634643,110.652386 0.0838624801,110.651806"
-										id="Fill-243"
-										fill="#C7D6E1"
-									/>
-									<path
-										d="M148.336365,106.470328 C150.833844,106.030484 151.603861,108.825061 151.603861,108.825061"
-										id="Stroke-314"
-										stroke="#7ED2F1"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M273.016028,75.7104306 C272.483921,75.3274529 272.088177,74.8261003 271.654716,74.3572428 C271.247947,73.8657548 270.820289,73.3916748 270.475029,72.8462217 C270.457621,72.8189491 270.465745,72.7829724 270.493017,72.7649841 C270.515067,72.7510576 270.543501,72.7539589 270.562069,72.7702065 C271.056459,73.1856792 271.46845,73.6725251 271.901911,74.1413827 C272.308099,74.6340312 272.751425,75.0941847 273.059548,75.6721329 C273.067091,75.6866396 273.061869,75.7046279 273.047362,75.7121714 C273.036917,75.7179741 273.024732,75.7168136 273.016028,75.7104306"
-										id="Fill-477"
-										fill="#C2F4FF"
-									/>
-									<path
-										d="M273.448096,75.5701796 C272.926434,75.1738557 272.543456,74.6626385 272.122181,74.1827558 C271.728758,73.6802427 271.313866,73.1951376 270.982532,72.6409805 C270.965704,72.6131276 270.974989,72.5771509 271.002842,72.5603231 C271.025472,72.5469769 271.053325,72.5504585 271.071894,72.5672863 C271.554678,72.995525 271.954483,73.4928157 272.375178,73.9732787 C272.768601,74.4757918 273.199741,74.9469704 273.492777,75.5336226 C273.49974,75.5481293 273.493937,75.5661177 273.479431,75.5730809 C273.468986,75.5783033 273.4568,75.5771428 273.448096,75.5701796"
-										id="Fill-478"
-										fill="#C2F4FF"
-									/>
-									<path
-										d="M273.835136,79.2348118 L273.835136,79.5668638 C273.835136,79.730138 274.064342,79.8622249 274.346353,79.8622249 L276.522363,79.8622249"
-										id="Fill-479"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M273.864323,79.2348118 L273.895078,79.4303625 C273.906103,79.4953527 273.907263,79.5719482 273.933956,79.5835536 C273.987921,79.6334568 274.098172,79.670594 274.204361,79.679298 C274.411517,79.6885823 274.710936,79.6694335 274.959291,79.6659518 C275.480373,79.6821994 276.001455,79.6700137 276.522537,79.7721411 C276.554451,79.7785241 276.575341,79.8098586 276.568958,79.8417734 C276.564316,79.8661447 276.545167,79.8841331 276.522537,79.888195 C276.001455,79.9909026 275.480373,79.978717 274.959291,79.9949645 C274.687725,79.9775564 274.463161,80.0036685 274.153297,79.9502838 C274.011712,79.9079241 273.847496,79.8464156 273.772061,79.66189 C273.737825,79.5545402 273.773801,79.4930316 273.775542,79.4303625 L273.806296,79.2348118 C273.808617,79.2191445 273.823704,79.2081194 273.839952,79.2104405 C273.852718,79.2121813 273.862583,79.2226262 273.864323,79.2348118"
-										id="Fill-480"
-										fill="#C8D7E2"
-									/>
-									<path
-										d="M271.816089,75.6900052 C273.28359,73.8087723 275.319755,75.5658275 277.457466,75.239136"
-										id="Fill-481"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M271.793168,75.6721329 C272.038042,75.2293874 272.474404,74.8643981 272.994326,74.7280348 C273.515407,74.603277 274.043452,74.6566617 274.538422,74.7431218 C275.527201,74.9462161 276.463755,75.2293874 277.448472,75.1818054 C277.480967,75.1800646 277.50824,75.2044359 277.509981,75.2369309 C277.511141,75.2647839 277.492573,75.2891552 277.46588,75.2966987 C276.970911,75.4226171 276.448088,75.4034682 275.949637,75.336157 C275.448864,75.2659444 274.962019,75.1469892 274.479815,75.0663318 C274.001093,74.9769703 273.517729,74.9131407 273.062217,75.0112062 C272.602064,75.0965058 272.20632,75.3715534 271.83901,75.7081096 C271.827404,75.7191347 271.808255,75.7179741 271.79723,75.7063687 C271.788526,75.6965042 271.787366,75.683158 271.793168,75.6721329"
-										id="Fill-482"
-										fill="#C8D7E2"
-									/>
-									<path
-										d="M278.063325,75.1273181 C277.929864,74.491343 276.590022,74.0224855 275.230451,74.6062363"
-										id="Fill-483"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M278.035008,75.1332368 C277.895744,74.9144753 277.699613,74.7902977 277.474468,74.6962941 C277.255707,74.5930062 277.009092,74.5425228 276.763058,74.5198923 C276.264607,74.4508402 275.766736,74.5512268 275.253778,74.6597372 C275.221863,74.6667004 275.189948,74.6475515 275.182405,74.6156367 C275.176602,74.5912654 275.188208,74.5663138 275.207356,74.5529676 C275.66809,74.2518079 276.240816,74.1502608 276.796714,74.1932007 C277.075243,74.2251155 277.350871,74.301711 277.605609,74.4357532 C277.847001,74.5715362 278.100579,74.8135085 278.091875,75.1210512 C278.091294,75.137879 278.077948,75.1506449 278.06112,75.1500646 C278.051256,75.1500646 278.042552,75.1448422 278.037329,75.1367185 L278.035008,75.1332368 Z"
-										id="Fill-484"
-										fill="#C8D7E2"
-									/>
-									<path
-										d="M279.007133,75.6900052 L279.007133,76.2627309 C279.007133,77.8834228 277.434604,79.1971523 275.494184,79.1971523 L275.191863,79.1971523 C273.252023,79.1971523 271.678913,77.8834228 271.678913,76.2627309 L271.678913,75.6900052 L278.088567,75.6900052"
-										id="Fill-485"
-										fill="#FFFFFF"
-									/>
-									<path
-										d="M279.036321,75.6900052 C279.05547,76.1652456 279.105373,76.6718207 278.934194,77.1511231 C278.771718,77.6234622 278.480423,78.0505404 278.10673,78.3847754 C277.362244,79.0631102 276.328204,79.3700726 275.342907,79.3480223 C274.358771,79.3747147 273.319509,79.0869012 272.551812,78.41611 C272.168254,78.0853565 271.859551,77.6582784 271.681989,77.1766549 C271.496883,76.6886485 271.521834,76.1577021 271.525316,75.6900052 C271.524736,75.6052858 271.594368,75.5362338 271.679087,75.5356535 C272.747363,75.5397154 273.815639,75.5449379 274.883914,75.5652473 C275.95219,75.5814948 277.019885,75.5919397 278.088161,75.6319782 C278.120076,75.6331388 278.145027,75.6598312 278.143867,75.691746 C278.142706,75.7225002 278.118335,75.7468715 278.088161,75.7480321 C277.019885,75.7880707 275.95219,75.7979352 274.883914,75.814763 C273.815639,75.8350725 272.747363,75.8402949 271.679087,75.8437765 L271.832859,75.6900052 C271.834019,76.183234 271.817192,76.6410664 271.983149,77.0664038 C272.142142,77.4917411 272.425314,77.8648542 272.766512,78.1694955 C273.460514,78.7840007 274.405192,79.0625299 275.342907,79.0462823 C276.280042,79.067172 277.230523,78.8072114 277.947736,78.2014104 C278.298799,77.8990901 278.599378,77.5271375 278.772879,77.0919356 C278.953342,76.657314 278.960886,76.1751102 278.978294,75.6900052 C278.978874,75.6737576 278.99222,75.661572 279.008468,75.6621522 C279.023555,75.6627325 279.035741,75.6749182 279.036321,75.6900052"
-										id="Fill-486"
-										fill="#C8D7E2"
-									/>
-								</g>
-								<g id="Page-1" transform="translate(27.000000, 38.000000)">
-									<g id="Group-252" transform="translate(2.281250, 0.705043)">
-										<path
-											d="M6.86117188,12.6475 C7.46585938,12.6365625 7.74632813,12.5529688 8.35023438,12.3514062 C10.8377344,11.5185938 12.0377344,9.37015625 11.9416406,6.950625 C11.9142969,6.26625 13.1010156,5.8053125 12.8650781,5.100625 C12.5596094,4.188125 10.9385156,4.25140625 10.1174219,2.49828125 C9.19710938,0.531875 6.76195312,0.02953125 4.48226563,0.79203125 C1.29085938,1.86 -0.431015625,5.31390625 0.636953125,8.5053125"
-											id="Fill-250"
-											fill="#FFFFFF"
-										/>
-										<path
-											d="M6.86117188,12.6475 C7.46585938,12.6365625 7.74632813,12.5529688 8.35023438,12.3514062 C10.8377344,11.5185938 12.0377344,9.37015625 11.9416406,6.950625 C11.9142969,6.26625 13.1010156,5.8053125 12.8650781,5.100625 C12.5596094,4.188125 10.9385156,4.25140625 10.1174219,2.49828125 C9.19710938,0.531875 6.76195312,0.02953125 4.48226563,0.79203125 C1.29085938,1.86 -0.431015625,5.31390625 0.636953125,8.5053125"
-											id="Stroke-251"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-									<path
-										d="M8.479375,5.73269886 C9.36765625,7.95301136 9.14578125,11.6537926 6.85125,12.5420739 C4.55671875,13.4303551 2.91765625,9.21082386 2.91765625,9.21082386 L8.479375,5.73269886 Z"
-										id="Fill-253"
-										fill="#4E748E"
-									/>
-									<g id="Group-256" transform="translate(0.718750, 37.423793)">
-										<path
-											d="M1.34664063,0.54 C-0.651796875,15.97125 1.02710938,48.9439062 1.02710938,48.9439062 L8.57476563,48.9439062 L9.24117187,21.3001562"
-											id="Fill-254"
-											fill="#005081"
-										/>
-										<path
-											d="M1.34664063,0.54 C-0.651796875,15.97125 1.02710938,48.9439062 1.02710938,48.9439062 L8.57476563,48.9439062 L9.24117187,21.3001562"
-											id="Stroke-255"
-											stroke="#005081"
-											stroke-width="0.735"
-										/>
-									</g>
-									<path
-										d="M1.74453125,85.5906676 C4.0390625,85.6648864 10.9945313,86.1086364 12.03125,86.4789489 C13.0671875,86.8492614 13.140625,86.9961364 11.8828125,87.2922301 C10.625,87.5883239 1.74453125,87.2922301 1.74453125,87.2922301 L1.74453125,85.5906676 Z"
-										id="Fill-257"
-										fill="#005081"
-									/>
-									<path
-										d="M1.74453125,85.5906676 C4.0390625,85.6648864 10.9945313,86.1086364 12.03125,86.4789489 C13.0671875,86.8492614 13.140625,86.9961364 11.8828125,87.2922301 C10.625,87.5883239 1.74453125,87.2922301 1.74453125,87.2922301 L1.74453125,85.5906676 Z"
-										id="Stroke-258"
-										stroke="#005081"
-										stroke-width="0.735"
-									/>
-									<path
-										d="M14.4005469,37.6308239 C14.4005469,37.6308239 15.2263281,49.3261364 15.6216406,61.3886364 C15.9528906,71.4753551 14.7427344,82.1214489 14.8442969,86.3675426 L6.62945312,86.3675426 C6.74039062,61.7214489 1.84351562,51.5081676 2.06539062,37.9636364"
-										id="Fill-259"
-										fill="#07A8DC"
-									/>
-									<path
-										d="M47.8352344,87.4720739 C46.8571094,87.8103551 43.7274219,88.4126989 42.7813281,88.7228551 C42.0086719,88.9767614 42.1922656,90.1150426 43.0055469,90.1150426 L51.4266406,90.1150426 C52.3969531,90.1150426 53.1422656,89.2556676 53.0055469,88.2947301 L52.8680469,87.3400426 L47.8352344,87.4720739 Z"
-										id="Fill-260"
-										fill="#07A8DC"
-									/>
-									<path
-										d="M2.06546875,37.9637145 L14.5115625,37.630902 L14.3264062,30.0066832 L11.4404687,28.4535582 L14.9553125,21.6441832 L12.6240625,19.2019957 L9.08578125,15.546527 C5.4459375,14.6988707 1.85453125,17.2105895 1.45765625,20.9262145 C0.5553125,29.377777 2.06546875,37.9637145 2.06546875,37.9637145 Z"
-										id="Fill-261"
-										fill="#07A8DC"
-									/>
-									<polygon
-										id="Fill-262"
-										fill="#7ED2F1"
-										points="10.2928906 27.8613707 10.8475781 68.2723082 16.0663281 68.2723082 15.2889844 30.3035582"
-									/>
-									<g id="Group-265" transform="translate(8.531250, 10.861293)">
-										<path
-											d="M0.318515625,15.3345312 C5.20289063,19.5532813 9.97710938,20.774375 11.9747656,19.1079688 C14.9591406,16.6196875 9.66226563,6.7821875 9.31070313,5.343125 C8.08960938,0.34703125 5.31460938,0.3478125 4.64820313,0.90328125 C3.98179688,1.45875 5.31460938,4.56578125 5.31460938,4.56578125 L6.53570313,11.2275"
-											id="Fill-263"
-											fill="#FFFFFF"
-										/>
-										<path
-											d="M0.318515625,15.3345312 C5.20289063,19.5532813 9.97710938,20.774375 11.9747656,19.1079688 C14.9591406,16.6196875 9.66226563,6.7821875 9.31070313,5.343125 C8.08960938,0.34703125 5.31460938,0.3478125 4.64820313,0.90328125 C3.98179688,1.45875 5.31460938,4.56578125 5.31460938,4.56578125 L6.53570313,11.2275"
-											id="Stroke-264"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-									<path
-										d="M12.1803906,19.424027 L16.7319531,23.7544957"
-										id="Stroke-266"
-										stroke="#005081"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M6.62929688,85.5906676 C8.92382813,85.6648864 15.8792969,86.1086364 16.9160156,86.4789489 C17.9519531,86.8492614 18.0261719,87.4406676 16.7675781,87.7367614 C15.5097656,88.0328551 6.62929688,87.7367614 6.62929688,87.7367614 L6.62929688,85.5906676 Z"
-										id="Fill-267"
-										fill="#07A8DC"
-									/>
-									<polygon
-										id="Stroke-268"
-										stroke="#005081"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-										points="10.1864844 13.5005114 9.51460938 15.5372301 8.99820313 13.4286364"
-									/>
-									<path
-										d="M12.4020312,37.410277 C8.62703125,39.0751207 1.30046875,38.1876207 1.30046875,38.1876207"
-										id="Stroke-269"
-										stroke="#7ED2F1"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M16.6954687,48.6579332 C15.4564062,48.885277 13.949375,48.804027 12.4548437,48.582152"
-										id="Stroke-270"
-										stroke="#FFFFFF"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M6.3065625,15.7466051 C8.270625,15.0794176 4.01125,12.8161364 3.75578125,14.3372301 C3.5003125,15.8583239 4.63078125,16.3153551 6.3065625,15.7466051 Z"
-										id="Stroke-271"
-										stroke="#7ED2F1"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M7.1846875,15.4273082 C7.1846875,15.4273082 9.626875,15.760902 10.8479688,16.9819957"
-										id="Stroke-272"
-										stroke="#7ED2F1"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M2.5215625,9.76598011 L14.7332813,1.44019886"
-										id="Stroke-273"
-										stroke="#7ED2F1"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M12.7339062,9.58207386 C13.2776562,9.62894886 13.7315625,9.44926136 13.7315625,9.44926136"
-										id="Stroke-274"
-										stroke="#005081"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M12.2071094,7.02512074 C12.0172656,7.06027699 11.8352344,6.93449574 11.8000781,6.74543324 L11.6938281,6.16418324 C11.6586719,5.97433949 11.7844531,5.79230824 11.9742969,5.75715199 C12.1633594,5.72277699 12.3453906,5.84855824 12.3805469,6.03762074 L12.4867969,6.61887074 C12.5219531,6.80871449 12.3969531,6.99074574 12.2071094,7.02512074"
-										id="Fill-275"
-										fill="#005081"
-									/>
-									<polygon
-										id="Fill-346"
-										fill="#C7D6E1"
-										points="49.7307813 29.1155895 44.9925 29.1155895 44.21125 18.3562145 49.7307813 18.3562145"
-									/>
-									<g id="Group-349" transform="translate(54.625000, 27.267543)">
-										<path
-											d="M0.046953125,0.341640625 C2.41570313,0.045546875 8.78054687,8.03929687 9.96492187,11.1478906 C11.1485156,14.2557031 3.45164062,16.0322656 3.45164062,16.0322656"
-											id="Fill-347"
-											fill="#C7D6E1"
-										/>
-										<path
-											d="M0.046953125,0.341640625 C2.41570313,0.045546875 8.78054687,8.03929687 9.96492187,11.1478906 C11.1485156,14.2557031 3.45164062,16.0322656 3.45164062,16.0322656"
-											id="Stroke-348"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-									<g id="Group-352" transform="translate(51.500000, 64.767543)">
-										<path
-											d="M11.4735156,1.43 C11.4735156,1.43 15.9563281,9.9753125 15.3102344,18.0565625 L14.5703906,20.2042187 L0.411796875,0.69875"
-											id="Fill-350"
-											fill="#C7D6E1"
-										/>
-										<path
-											d="M11.4735156,1.43 C11.4735156,1.43 15.9563281,9.9753125 15.3102344,18.0565625 L14.5703906,20.2042187 L0.411796875,0.69875"
-											id="Stroke-351"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-									<g id="Group-355" transform="translate(45.250000, 62.423793)">
-										<path
-											d="M0.68859375,2.2684375 C0.68859375,2.2684375 1.6878125,22.4739062 2.79796875,25.9153125 L7.01671875,25.804375 C7.23859375,18.6989062 9.68078125,9.595 9.90265625,0.491875"
-											id="Fill-353"
-											fill="#C7D6E1"
-										/>
-										<path
-											d="M0.68859375,2.2684375 C0.68859375,2.2684375 1.6878125,22.4739062 2.79796875,25.9153125 L7.01671875,25.804375 C7.23859375,18.6989062 9.68078125,9.595 9.90265625,0.491875"
-											id="Stroke-354"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-									<g id="Group-358" transform="translate(43.687500, 23.361293)">
-										<path
-											d="M0.566875,0.18984375 C1.59109375,0.4375 2.73328125,0.1015625 2.73328125,0.1015625 C2.80046875,0.09921875 2.83328125,0.17265625 2.78171875,0.21015625 L0.995,2.109375 L0.523125,0.07578125"
-											id="Fill-356"
-											fill="#005081"
-										/>
-										<path
-											d="M0.566875,0.18984375 C1.59109375,0.4375 2.73328125,0.1015625 2.73328125,0.1015625 C2.80046875,0.09921875 2.83328125,0.17265625 2.78171875,0.21015625 L0.995,2.109375 L0.523125,0.07578125"
-											id="Stroke-357"
-											stroke="#005081"
-											stroke-width="0.876"
-										/>
-									</g>
-									<g id="Group-361" transform="translate(38.218750, 10.861293)">
-										<path
-											d="M7.2540625,0.440859375 C4.85328125,0.325234375 2.64625,1.46976562 2.293125,3.61273437 C1.9790625,5.52210938 0.40171875,5.90179687 0.355625,6.86273437 C0.3196875,7.60492188 1.586875,7.72601563 1.74625,8.39320313 C2.30953125,10.7471094 4.04859375,12.4900781 6.668125,12.6158594 C7.30484375,12.6463281 7.59625,12.6502344 8.1821875,12.4971094"
-											id="Fill-359"
-											fill="#C7D6E1"
-										/>
-										<path
-											d="M7.2540625,0.440859375 C4.85328125,0.325234375 2.64625,1.46976562 2.293125,3.61273437 C1.9790625,5.52210938 0.40171875,5.90179687 0.355625,6.86273437 C0.3196875,7.60492188 1.586875,7.72601563 1.74625,8.39320313 C2.30953125,10.7471094 4.04859375,12.4900781 6.668125,12.6158594 C7.30484375,12.6463281 7.59625,12.6502344 8.1821875,12.4971094"
-											id="Stroke-360"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-									<path
-										d="M41.6928125,19.7600426 C41.3709375,20.5116051 40.5654688,20.8342614 40.5654688,20.8342614"
-										id="Stroke-362"
-										stroke="#005081"
-										stroke-linecap="round"
-										stroke-linejoin="round"
-									/>
-									<path
-										d="M41.34375,17.3434801 C41.5367188,17.3434801 41.6929688,17.1872301 41.6929688,16.9942614 L41.6929688,16.4036364 C41.6929688,16.2106676 41.5367188,16.0544176 41.34375,16.0544176 C41.1507813,16.0544176 40.9945313,16.2106676 40.9945313,16.4036364 L40.9945313,16.9942614 C40.9945313,17.1872301 41.1507813,17.3434801 41.34375,17.3434801"
-										id="Fill-363"
-										fill="#005081"
-									/>
-									<path
-										d="M44.4954688,26.1657457 C51.0267188,24.0938707 53.8290625,25.5993395 56.3189063,29.385277 C60.37125,35.5469957 58.0579688,37.0829332 59.5384375,42.7079332 C61.018125,48.3329332 64.4790625,50.7594957 63.8126563,66.8016832 L45.0501563,66.8016832 C45.0501563,66.8016832 43.274375,48.3141832 43.8290625,42.9298082"
-										id="Fill-364"
-										fill="#07A8DC"
-									/>
-									<path
-										d="M63.903125,82.7216832 C63.903125,82.7216832 68.259375,90.3732457 69.7226563,90.055277 C70.6023438,89.8638707 69.2070313,86.142777 68.7429688,84.3982457 C68.4195313,83.1826207 68.1492188,82.383402 66.8976563,82.5099645 L63.903125,82.7216832 Z"
-										id="Fill-365"
-										fill="#07A8DC"
-									/>
-									<g id="Group-368" transform="translate(31.187500, 23.361293)">
-										<path
-											d="M9.49671875,9.57703125 C9.49671875,9.57703125 8.2084375,8.6309375 7.21703125,6.8871875 C5.71703125,4.2465625 1.92796875,-0.27453125 0.985,0.91765625 C0.04125,2.1090625 1.71,6.27546875 3.13109375,6.98640625 C4.16703125,10.316875 9.53265625,21.598125 12.7521875,20.7035938 C15.9717188,19.8090625 18.5248438,2.7075 14.0842188,2.9309375 C9.64359375,3.154375 9.31078125,12.9801562 9.31078125,12.9801562"
-											id="Fill-366"
-											fill="#C7D6E1"
-										/>
-										<path
-											d="M9.49671875,9.57703125 C9.49671875,9.57703125 8.2084375,8.6309375 7.21703125,6.8871875 C5.71703125,4.2465625 1.92796875,-0.27453125 0.985,0.91765625 C0.04125,2.1090625 1.71,6.27546875 3.13109375,6.98640625 C4.16703125,10.316875 9.53265625,21.598125 12.7521875,20.7035938 C15.9717188,19.8090625 18.5248438,2.7075 14.0842188,2.9309375 C9.64359375,3.154375 9.31078125,12.9801562 9.31078125,12.9801562"
-											id="Stroke-367"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-									<path
-										d="M56.3782812,22.0091832 C55.9079687,21.2466832 55.3345312,20.6044957 54.7071875,20.0982457 C54.1392187,19.6388707 53.75875,18.999027 53.649375,18.277152 C53.4978125,17.2779332 53.1485937,16.3099645 52.6735937,15.6193395 C52.015,14.661527 50.8821875,13.9232457 49.7001562,13.6193395 C48.6267187,11.714652 47.6579687,11.4349645 46.30875,11.030277 C44.0884375,10.3638707 42.0157812,11.6224645 42.0157812,11.6224645 C41.7204687,14.0849645 42.6892187,16.852152 44.99625,17.0076207 C44.5954687,18.8849645 45.6798437,21.4568395 47.7728125,22.0701207 C47.549375,23.4857457 47.8829687,25.1560582 48.82125,26.6748082 C50.5415625,29.4623082 53.6282812,30.677152 55.7157812,29.389652 C57.8025,28.1005895 58.099375,24.7966832 56.3782812,22.0091832"
-										id="Fill-369"
-										fill="#005081"
-									/>
-									<path
-										d="M33.6491406,27.9594176 L31.2069531,21.0125426 C31.0694531,20.6219176 31.2741406,20.1945739 31.6647656,20.0570739 L35.0163281,18.8781676 C35.4069531,18.7414489 35.8342969,18.9469176 35.9717969,19.3367614 L38.4147656,26.2836364 C38.5514844,26.6742614 38.3467969,27.1023864 37.9561719,27.2398864 L34.6046094,28.4180114 C34.2147656,28.5555114 33.7866406,28.3500426 33.6491406,27.9594176"
-										id="Fill-452"
-										fill="#00A9DE"
-									/>
-									<polygon
-										id="Fill-453"
-										fill="#FFFFFF"
-										points="37.5626562 25.8181676 34.0220312 27.0626989 31.8954687 21.0150426 35.4360937 19.7705114"
-									/>
-									<polygon
-										id="Fill-454"
-										fill="#004F84"
-										points="35.4290625 19.750902 31.8884375 20.9962145 32.1845312 21.8368395 35.724375 20.5923082"
-									/>
-									<g id="Group-457" transform="translate(34.312500, 26.486293)">
-										<path
-											d="M1.62320313,1.54515625 C1.62320313,1.54515625 0.972421875,-0.0275 0.716953125,0.21078125 C0.164609375,0.7271875 0.831796875,2.88265625 0.831796875,2.88265625"
-											id="Fill-455"
-											fill="#C7D6E1"
-										/>
-										<path
-											d="M1.62320313,1.54515625 C1.62320313,1.54515625 0.972421875,-0.0275 0.716953125,0.21078125 C0.164609375,0.7271875 0.831796875,2.88265625 0.831796875,2.88265625"
-											id="Stroke-456"
-											stroke="#005081"
-											stroke-linecap="round"
-											stroke-linejoin="round"
-										/>
-									</g>
-								</g>
-							</g>
-						</g>
-					</svg>
-				</Card>
-
-				<CTACard
-					headerText={ translate( 'Physical Location or Service Area' ) }
-					mainText={ translate(
-						'My business has a physical location custmers can visit, ' +
-							'or provides goods and services to local customers, or both.'
-					) }
-					buttonText={ translate( 'Create My Listing' ) }
-					buttonIcon="external"
-					buttonPrimary={ true }
-				/>
-
-				<CTACard
-					headerText={ translate( 'Online Only' ) }
-					mainText={ translate(
-						"Don't provide in-person services? Learn more about reaching your customers online."
-					) }
-					buttonText={ translate( 'Optimize Your SEO' ) }
-					buttonIcon="external"
-				/>
-			</div>
-		);
-	}
-}
-
-export default localize( SelectBusinessType );
